// src/entities/interfaces/user.interface.ts
// export interface UserI {
//   id: number;
//   userName?: string;
//   email: string;
//   password?: string;
//   firstName?: string;
//   lastName?: string;
//   birthday?: Date;
//   isActive?: boolean;
//   points?: number;
//   created: Date;
//   updated: Date;
// }

export interface UserI {
  id: number;
  userName?: string;
  email: string;
  password?: string;
  created: Date;
  updated: Date;
  firstName?: string;
  lastName?: string;
}